--- src/components/Cards.js
@@ -1,16 +1,0 @@
-import React from "react";
-import ApplicantLabel from "./ApplicantLabel";
-import IconContainer from "./IconContainer";
-
-const Cards = ({ students }) => {
-  return students
-    .map(student => (
-      <div className="card">
-        <ApplicantLabel student={student} />
-        <IconContainer student={student} />
-      </div>
-    ))
-    .reverse();
-};
-
-export default Cards;
