+++ src/components/Column.js
@@ -1,40 +1,15 @@
 import React from "react";
-import Cards from "./Cards";
+import Card from "./Card";
 
-const currentStep = student => {
-  let currentStepNum = 0;
-  // Find the approved steps.
-  const steps = student.steps.map(step => step.number).sort();
-  const approvedSteps = student.steps.filter(
-    step => step.status === "Approved"
-  );
-
-  // Find out on which step student is.
-  // if (approvedSteps.length > 0) {
-  //   approvedSteps.forEach(step => {
-  //     if (step.number > currentStepNum) {
-  //       currentStepNum = step.number;
-  //     }
-  //   });
-  // }
-  if (steps.length !== 0) {
-    currentStepNum = steps[steps.length - 1];
-  }
-
-  return currentStepNum;
-};
-
-const Column = ({ students, step }) => {
-  // Find the students which is allowed to be added to column.
-  const allowedStudents = students.filter(
-    student => currentStep(student) === step.number
-  );
-  console.log("columnComp", { allowedStudents, step });
+const Column = ({ allowedStudents, step }) => {
+  const card = allowedStudents
+    .map(student => <Card key={student._id} student={student} />)
+    .reverse();
 
   return (
     <div className="column">
-      <h3>{step.header}</h3>
-      <Cards key={students._id} students={allowedStudents} />
+      <h3>{`${step.header} (${allowedStudents.length})`}</h3>
+      {card}
     </div>
   );
 };
